# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  page.CreatePageRequest:
    properties:
      content:
        type: string
      slug:
        type: string
      status:
        type: string
      template:
        type: string
      title:
        type: string
    required:
    - slug
    - title
    type: object
  page.UpdatePageRequest:
    properties:
      content:
        type: string
      slug:
        type: string
      status:
        type: string
      template:
        type: string
      title:
        type: string
    type: object
  response.Body:
    properties:
      data: {}
      error:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
  seo.CreateSEORequest:
    properties:
      canonical_url:
        type: string
      description:
        type: string
      keywords:
        type: string
      og_description:
        type: string
      og_image:
        type: string
      og_title:
        type: string
      og_type:
        type: string
      robots:
        type: string
      slug:
        type: string
      structured_data:
        type: string
      title:
        type: string
      twitter_card:
        type: string
    required:
    - slug
    - title
    type: object
  seo.UpdateSEORequest:
    properties:
      canonical_url:
        type: string
      description:
        type: string
      keywords:
        type: string
      og_description:
        type: string
      og_image:
        type: string
      og_title:
        type: string
      og_type:
        type: string
      robots:
        type: string
      slug:
        type: string
      structured_data:
        type: string
      title:
        type: string
      twitter_card:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Modular Go + Gin Backend API (REST & GraphQL)
  title: CodebaseGo API
  version: "1.0"
paths:
...
  /pages:
    get:
      parameters:
      - description: Opaque cursor for pagination
        in: query
        name: cursor
        type: string
      - default: 10
        description: Items per page for cursor pagination
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page number for offset pagination
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page for offset pagination
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: List pages
      tags:
      - pages
    post:
      consumes:
      - application/json
      parameters:
      - description: Page payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/page.CreatePageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Create page
      tags:
      - pages
  /pages/{id}:
    delete:
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Delete page
      tags:
      - pages
    get:
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      summary: Get page by ID
      tags:
      - pages
    put:
      consumes:
      - application/json
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: string
      - description: Page update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/page.UpdatePageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Update page
      tags:
      - pages
  /pages/by-slug:
    get:
      parameters:
      - description: Page slug (e.g. 'home', 'about-us')
        in: query
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      summary: Get page by slug
      tags:
      - pages
...
  /seo:
    get:
      parameters:
      - description: Opaque cursor for pagination
        in: query
        name: cursor
        type: string
      - default: 10
        description: Items per page for cursor pagination
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page number for offset pagination
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page for offset pagination
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: List SEO records
      tags:
      - seo
    post:
      consumes:
      - application/json
      parameters:
      - description: SEO Metadata payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/seo.CreateSEORequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Create SEO record
      tags:
      - seo
  /seo/{id}:
    delete:
      parameters:
      - description: SEO Record ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Delete SEO record
      tags:
      - seo
    get:
      parameters:
      - description: SEO Record ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      summary: Get SEO record by ID
      tags:
      - seo
    put:
      consumes:
      - application/json
      parameters:
      - description: SEO Record ID
        in: path
        name: id
        required: true
        type: string
      - description: SEO Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/seo.UpdateSEORequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Update SEO record
      tags:
      - seo
  /seo/by-slug:
    get:
      parameters:
      - description: Page slug (e.g. 'home', 'about-us', 'products/shoes')
        in: query
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      summary: Get SEO record by slug
      tags:
      - seo
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"